+++ docs/.vitepress/config/plugin.ts
@@ -1,5 +1,5 @@
-import { demoPreviewPlugin } from '@movk-repo/demo-preview-plugin'
 import type MarkdownIt from 'markdown-it'
+import { demoPreviewPlugin } from '@movk-repo/demo-preview-plugin'
 
 export function mdPlugin(md: MarkdownIt) {
   md.use(demoPreviewPlugin)