+++ internal/built-utils/tsup.config.ts
@@ -1,0 +1,11 @@
+import { defineConfig } from 'tsup'
+
+export default defineConfig({
+  entryPoints: [
+    'src/index.ts',
+  ],
+  clean: true,
+  format: ['esm'],
+  dts: true,
+  shims: true,
+})