+++ packages/components/index.ts
@@ -1,2 +1,0 @@
-export * from './client-only/client-only'
-export * from './message'