+++ packages/demo-preview/container/src/index.ts
@@ -1,0 +1,4 @@
+import DemoPreviewContainer from './Container.vue'
+import './index.scss'
+
+export { DemoPreviewContainer }