+++ packages/auto-nav-sidebar/CHANGELOG.md
@@ -1,5 +1,15 @@
-## Features
+# 1.0.2 (2024-09-23)
 
-### 1.0.0
+Bug Fixes
 
-- Initial release
+- 修复了README.md中的错误
+
+## 1.0.1 (2024-09-23)
+
+Performance Improvements
+
+- 优化了打包文件
+
+## 1.0.0 (2024-09-1)
+
+Initial release